+++ scripts/git.sh
@@ -2,6 +2,6 @@
 
 echo -e "Updating your Git setup and configuration..."
 
-cp -rf .gitignore ~/.gitignore
-cp -rf .gitconfig ~/.gitconfig
-cp -rf .git-prompt.sh ~/.git-prompt.sh
+[ -f ~/.gitignore ] && cp -rf .gitignore ~/.gitignore || cp .gitignore ~/.gitignore
+[ -f ~/.gitconfig ] && cp -rf .gitconfig ~/.gitconfig || cp .gitconfig ~/.gitconfig
+[ -f ~/.git-prompt.sh ] && cp -rf .git-prompt.sh ~/.git-prompt.sh || cp .git-prompt.sh ~/.git-prompt.sh
